feat(playground): centralize profile menu data
