+++ playground/src/layouts/UserLayout.vue
@@ -12,6 +12,15 @@
         <template #navLogo>
             <img src="/atlas.png" alt="Atlas" class="h-8 w-8 rounded-full" />
         </template>
+        <template #navActions>
+            <ProfileMenu
+                :user="user"
+                :items="profileMenuItems"
+                :avatar-only="true"
+                headerLink="/"
+                :linkComponent="Link"
+            />
+        </template>
         <template #headerTitle>
             <div class="pr-2">
                 <Button
@@ -53,11 +62,13 @@
 import { computed } from 'vue';
 import { useRoute } from 'vue-router';
 import LayoutApp from '@atlas/ui/components/App/Index.vue';
+import ProfileMenu from '@atlas/ui/components/App/Nav/ProfileMenu.vue';
 import { Button, useModal } from '@atlas/ui';
 import UserModals from '../components/UserModals.vue';
 import Link from '../components/RouterLink.vue';
 import { sideBarItems } from '../sideBarItems';
 import { useSettings } from '../composables/useSettings';
+import { useProfileMenu } from '../composables/useProfileMenu';
 
 const props = defineProps({
     item: {
@@ -70,4 +81,5 @@ const { open } = useModal();
 const route = useRoute();
 const { topNav } = useSettings();
 const topBarItems = computed(() => sideBarItems.flatMap((section) => section.children));
+const { user, profileMenuItems } = useProfileMenu();
 </script>
